+++ gulp/clean.js
@@ -1,0 +1,4 @@
+const del = require('del');
+
+exports.cleanDist = () => del('client-dist');
+exports.cleanTmp = () => del('tmp');
