+++ package.json
@@ -5,11 +5,6 @@
   "main": "client-dist/app.js",
   "scripts": {
     "test": "gulp test",
-    "test:unit": "gulp test:unit",
-    "test:api": "gulp test:api:integration",
-    "pretest": "dropdb moolah_test_db -U postgres || true && createdb moolah_test_db -U postgres",
-    "pretest:api": "dropdb moolah_test_db -U postgres || true && createdb moolah_test_db -U postgres",
-    "precoverage": "dropdb moolah_test_db -U postgres || true && createdb moolah_test_db -U postgres",
     "lint": "gulp lint",
     "test-browser": "gulp test-browser",
     "watch-tests": "gulp watch-tests",
